gain and gain ratio
fixed the function to properly spit out a list of gain and gain ratios to each attribute in list form.
attribute list are in order from 0 to 13
sub_index 0 is gain
sub_index 1 is gain ratio
Note: copied over gain function from entropy file to main file for it to properly work.